requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
python-dateutil>=2.8.2
pytest>=7.0.0
pytest-django>=4.5.2
django-cors-headers>=3.12.0